refactor: 바인딩 어댑터를 한 파일로 관리
