+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketentry/TicketEntryBindingAdapter.kt
@@ -1,9 +1,24 @@
 package com.festago.festago.presentation.ui.ticketentry
 
 import android.view.ViewGroup
+import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.festago.festago.R
 import com.festago.festago.presentation.model.TicketStateUiModel
+import com.google.zxing.BarcodeFormat
+import com.journeyapps.barcodescanner.BarcodeEncoder
+
+@BindingAdapter("qrContent")
+fun setQrCodeImage(imageView: ImageView, content: String) {
+    val barcodeEncoder = BarcodeEncoder()
+    val bitmap = barcodeEncoder.encodeBitmap(
+        content,
+        BarcodeFormat.QR_CODE,
+        200,
+        200,
+    )
+    imageView.setImageBitmap(bitmap)
+}
 
 @BindingAdapter("ticketBackground")
 fun setTicketBackgroundByState(viewGroup: ViewGroup, state: TicketStateUiModel) {
